+++ src/lib/mrvApp/mrvApp.cpp
@@ -18,14 +18,6 @@
 // #include <tlPlay/ViewportModel.h>
 // #include <tlPlay/Util.h>
 
-#ifdef TLRENDER_BMD
-#    include <tlDevice/BMDOutputDevice.h>
-#endif
-
-#ifdef TLRENDER_NDI
-#    include <tlDevice/NDIOutputDevice.h>
-#endif
-
 #ifdef MRV2_PYBIND11
 #    include <pybind11/embed.h>
 namespace py = pybind11;
@@ -105,6 +97,18 @@ namespace py = pybind11;
 #endif
 
 #include "mrvFl/mrvIO.h"
+
+#if defined(TLRENDER_NDI) || defined(TLRENDER_BMD)
+#    include <tlDevice/DevicesModel.h>
+#endif
+
+#ifdef TLRENDER_BMD
+#    include <tlDevice/BMD/BMDOutputDevice.h>
+#endif
+
+#ifdef TLRENDER_NDI
+#    include <tlDevice/NDI/NDIOutputDevice.h>
+#endif
 
 #ifdef TLRENDER_NDI
 #    include <Processing.NDI.Lib.h>
@@ -358,7 +362,7 @@ namespace mrv
                     p.options.otioEditMode, {"-editMode", "-e"},
                     _("OpenTimelineIO Edit mode.")),
                 app::CmdLineFlagOption::create(
-                    p.options.singleImages, {"--single", "-single", "-s"},
+                    p.options.singleImages, {"-single", "-s"},
                     _("Load the images as still images not sequences.")),
                 app::CmdLineValueOption<double>::create(
                     p.options.speed, {"-speed"}, _("Playback speed.")),
@@ -478,8 +482,7 @@ namespace mrv
 #endif
 
                 app::CmdLineFlagOption::create(
-                    p.options.displayVersion,
-                    {"-version", "--version", "-v", "--v"},
+                    p.options.displayVersion, {"-version", "-v"},
                     _("Return the version and exit."))});
 
         const int exitCode = getExit();
@@ -525,6 +528,9 @@ namespace mrv
         Fl::lock(); // needed for NDI and multithreaded logging
 
         DBG;
+        // Create the Settings
+        p.settings = new SettingsObject();
+        
         // Create the interface.
         ui = new ViewerUI();
         if (!ui)
@@ -532,9 +538,6 @@ namespace mrv
             throw std::runtime_error(_("Cannot create window"));
         }
         DBG;
-
-        // Create the Settings
-        p.settings = new SettingsObject();
 
         // Classes used to handle network connections
 #ifdef MRV2_NETWORK
@@ -670,7 +673,7 @@ namespace mrv
         p.settings->setDefaultValue(
             "BMD/DeviceEnabled", bmdDevicesModelData.deviceEnabled);
         const auto i = bmdDevicesModelData.boolOptions.find(
-            bmd::Option::_444SDIVideoOutput);
+            device::Option::_444SDIVideoOutput);
         p.settings->setDefaultValue(
             "BMD/444SDIVideoOutput",
             i != bmdDevicesModelData.boolOptions.end() ? i->second : false);
@@ -682,6 +685,26 @@ namespace mrv
         if (!NDIlib_initialize())
             throw std::runtime_error(_("Could not initialize NDI"));
 #endif
+
+#if defined(TLRENDER_NDI)
+        device::DevicesModelData devicesModelData;
+        p.settings->setDefaultValue(
+            "NDI/DeviceIndex", devicesModelData.deviceIndex);
+        p.settings->setDefaultValue(
+            "NDI/DisplayModeIndex", devicesModelData.displayModeIndex);
+        p.settings->setDefaultValue(
+            "NDI/PixelTypeIndex", devicesModelData.pixelTypeIndex);
+        p.settings->setDefaultValue(
+            "NDI/DeviceEnabled", devicesModelData.deviceEnabled);
+        const auto i = devicesModelData.boolOptions.find(
+            device::Option::_444SDIVideoOutput);
+        p.settings->setDefaultValue(
+            "NDI/444SDIVideoOutput",
+            i != devicesModelData.boolOptions.end() ? i->second : false);
+        // p.settings->setDefaultValue("NDI/HDRMode", devicesModelData.hdrMode);
+        // p.settings->setDefaultValue("NDI/HDRData", devicesModelData.hdrData)
+        ;
+#endif // TLRENDER_NDI
 
         p.volume = p.settings->getValue<float>("Audio/Volume");
         p.mute = p.settings->getValue<bool>("Audio/Mute");
@@ -723,6 +746,72 @@ namespace mrv
                         player->setCompareTime(value);
                     }
                 });
+
+#if defined(TLRENDER_BMD)
+        p.devicesObserver =
+            observer::ValueObserver<device::DevicesModelData>::create(
+                p.devicesModel->observeData(),
+                [this](const device::DevicesModelData& value)
+                {
+                    TLRENDER_P();
+
+                    device::DeviceConfig config;
+                    config.deviceIndex = value.deviceIndex - 1;
+                    config.displayModeIndex = value.displayModeIndex - 1;
+                    config.pixelType =
+                        value.pixelTypeIndex >= 0 &&
+                                value.pixelTypeIndex < value.pixelTypes.size()
+                            ? value.pixelTypes[value.pixelTypeIndex]
+                            : device::PixelType::None;
+                    config.boolOptions = value.boolOptions;
+                    p.outputDevice->setConfig(config);
+                    p.outputDevice->setEnabled(value.deviceEnabled);
+                    p.putputVideoLevels = value.videoLevels;
+                    timeline::DisplayOptions displayOptions =
+                        p.viewportModel->getDisplayOptions();
+                    displayOptions.videoLevels = p.bmdOutputVideoLevels;
+                    p.outputDevice->setDisplayOptions({displayOptions});
+                    p.outputDevice->setHDR(value.hdrMode, value.hdrData);
+
+                    p.settings->setValue("NDI/DeviceIndex", value.deviceIndex);
+                    p.settings->setValue(
+                        "NDI/DisplayModeIndex", value.displayModeIndex);
+                    p.settings->setValue(
+                        "NDI/PixelTypeIndex", value.pixelTypeIndex);
+                    p.settings->setValue(
+                        "NDI/DeviceEnabled", value.deviceEnabled);
+                    const auto i = value.boolOptions.find(
+                        device::Option::_444SDIVideoOutput);
+                    p.settings->setValue(
+                        "NDI/444SDIVideoOutput",
+                        i != value.boolOptions.end() ? i->second : false);
+                    // p.settings->setValue("NDI/HDRMode", value.hdrMode);
+                    // p.settings->setValue("NDI/HDRData", value.hdrData);
+                });
+
+        p.bmdActiveObserver = observer::ValueObserver<bool>::create(
+            p.bmdOutputDevice->observeActive(),
+            [this](bool value)
+            {
+                _p->bmdDeviceActive = value;
+                _audioUpdate();
+            });
+        p.bmdSizeObserver = observer::ValueObserver<math::Size2i>::create(
+            p.bmdOutputDevice->observeSize(),
+            [this](const math::Size2i& value)
+            {
+                // std::cout << "output device size: " << value << std::endl;
+            });
+        p.bmdFrameRateObserver =
+            observer::ValueObserver<otime::RationalTime>::create(
+                p.bmdOutputDevice->observeFrameRate(),
+                [this](const otime::RationalTime& value)
+                {
+                    // std::cout << "output device frame rate: " << value <<
+                    // std::endl;
+                });
+
+#endif // TLRENDER_BMD || TLRENDER_NDI
 
         p.logObserver = observer::ListObserver<log::Item>::create(
             ui->app->getContext()->getLogSystem()->observeLog(),
@@ -797,7 +886,9 @@ namespace mrv
         // We make sure the UI is visible when we feed a filename.
         // This is needed to avoid an issue with Wayland not properly
         // refreshing the play buttons.
+#ifdef MRV2_PYBIND11
         if (p.options.pythonScript.empty())
+#endif
         {
             ui->uiMain->show();
             ui->uiView->take_focus();
@@ -1327,20 +1418,66 @@ namespace mrv
 
     void App::_startOutputDeviceTimer()
     {
+        TLRENDER_P();
+
+#    if defined(TLRENDER_NDI) || defined(TLRENDER_BMD)
+        if (p.outputDevice)
+        {
+            p.outputDevice->setPlayer(p.player ? p.player->player() : nullptr);
+            p.outputDevice->setEnabled(true);
+        }
+#    endif
         Fl::add_timeout(kTimeout, (Fl_Timeout_Handler)_timer_update_cb, this);
     }
 
     void App::_stopOutputDeviceTimer()
     {
+        TLRENDER_P();
+
+#    if defined(TLRENDER_NDI) || defined(TLRENDER_BMD)
+        if (p.outputDevice)
+            p.outputDevice->setEnabled(false);
+#    endif
         Fl::remove_timeout((Fl_Timeout_Handler)_timer_update_cb, this);
     }
 #endif // TLRENDER_BMD || TLRENDER_NDI
 
 #ifdef TLRENDER_NDI
-    void App::beginNDIOutputStream(const device::DeviceConfig& options)
+    void App::beginNDIOutputStream()
     {
         TLRENDER_P();
+        device::DeviceConfig config;
         p.outputDevice = ndi::OutputDevice::create(_context);
+        p.outputDevice->setConfig(config);
+        p.devicesModel = device::DevicesModel::create(_context);
+        p.devicesModel->setDeviceIndex(
+            p.settings->getValue<int>("NDI/DeviceIndex"));
+        p.devicesModel->setDisplayModeIndex(
+            p.settings->getValue<int>("NDI/DisplayModeIndex"));
+        p.devicesModel->setPixelTypeIndex(
+            p.settings->getValue<int>("NDI/PixelTypeIndex"));
+        p.devicesModel->setDeviceEnabled(
+            p.settings->getValue<bool>("NDI/DeviceEnabled"));
+        device::BoolOptions deviceBoolOptions;
+        deviceBoolOptions[device::Option::_444SDIVideoOutput] =
+            p.settings->getValue<bool>("NDI/444SDIVideoOutput");
+        p.devicesModel->setBoolOptions(deviceBoolOptions);
+        p.devicesModel->setHDRMode(static_cast<device::HDRMode>(
+            p.settings->getValue<int>("NDI/HDRMode")));
+        std::string s = p.settings->getValue<std::string>("NDI/HDRData");
+        if (!s.empty())
+        {
+            auto json = nlohmann::json::parse(s);
+            image::HDRData hdrData;
+            try
+            {
+                from_json(json, hdrData);
+            }
+            catch (const std::exception&)
+            {
+            }
+            p.devicesModel->setHDRData(hdrData);
+        }
         _startOutputDeviceTimer();
     }
 
@@ -1348,21 +1485,52 @@ namespace mrv
     {
         _stopOutputDeviceTimer();
         _p->outputDevice.reset();
+        _p->devicesModel.reset();
     }
 #endif
 
 #ifdef TLRENDER_BMD
-    void App::beginBMDOutputStream(const device::DeviceConfig& options)
+    void App::beginBMDOutputStream()
     {
         TLRENDER_P();
         p.outputDevice = bmd::OutputDevice::create(_context);
+        p.devicesModel = device::DevicesModel::create(_context);
+        p.devicesModel->setDeviceIndex(
+            p.settings->getValue<int>("BMD/DeviceIndex"));
+        p.devicesModel->setDisplayModeIndex(
+            p.settings->getValue<int>("BMD/DisplayModeIndex"));
+        p.devicesModel->setPixelTypeIndex(
+            p.settings->getValue<int>("BMD/PixelTypeIndex"));
+        p.devicesModel->setDeviceEnabled(
+            p.settings->getValue<bool>("BMD/DeviceEnabled"));
+        device::BoolOptions deviceBoolOptions;
+        deviceBoolOptions[bmd::Option::_444SDIVideoOutput] =
+            p.settings->getValue<bool>("BMD/444SDIVideoOutput");
+        p.devicesModel->setBoolOptions(deviceBoolOptions);
+        p.devicesModel->setHDRMode(static_cast<device::HDRMode>(
+            p.settings->getValue<int>("BMD/HDRMode")));
+        std::string s = p.settings->getValue<std::string>("BMD/HDRData");
+        if (!s.empty())
+        {
+            auto json = nlohmann::json::parse(s);
+            image::HDRData hdrData;
+            try
+            {
+                from_json(json, hdrData);
+            }
+            catch (const std::exception&)
+            {
+            }
+            p.devicesModel->setHDRData(hdrData);
+        }
         _startOutputDeviceTimer();
     }
 
     void App::endBMDOutputStream()
     {
         _stopOutputDeviceTimer();
-        p.outputDevice.reset();
+        _p->outputDevice.reset();
+        _p->devicesModel.reset();
     }
 #endif