+++ src/lib/mrvGL/mrvTimelineViewportPrivate.h
@@ -24,7 +24,7 @@ namespace mrv
         static std::map<
             std::string, std::string, string::CaseInsensitiveCompare>
             tagData;
-        static timeline::BackgroundOptions backgroundOptions;
+        std::shared_ptr<observer::Value<timeline::BackgroundOptions> > backgroundOptions;
         static float rotation;
         static bool resizeWindow;
 
@@ -180,6 +180,9 @@ namespace mrv
         // Observers
         std::shared_ptr<observer::ListObserver<timeline::VideoData> >
             videoDataObserver;
+
+        //! Overlay
+        std::shared_ptr<image::Image> overlayImage;
     };
 
 } // namespace mrv